Added multiple waves and wave types
Now working on more opitmization

+++ src/main/presets/EarthPreset.kt
@@ -7,16 +7,16 @@ import main.geometry.Vec2
 object EarthPreset {
      const val TOTAL_RADIUS: Int = 6370
 
-     const val CRUST_DV: Double      = 17.0 / 300.0
-     const val LITHO_DV: Double      = 17.0 / 700.0
-     const val ASTHENO_DV: Double    = 7.0 / 1500.0
-     const val MANTLE_DV: Double     = 1.0 / 600.0
-     const val OUT_DV: Double        = 1.0 / 2210.0
+     const val CRUST_DV: Float      = 17.0f / 300.0f
+     const val LITHO_DV: Float      = 17.0f / 700.0f
+     const val ASTHENO_DV: Float    = 7.0f / 1500.0f
+     const val MANTLE_DV: Float     = 1.0f / 600.0f
+     const val OUT_DV: Float        = 1.0f / 2210.0f
 
-     const val CRUST_VI: Double      = 5.3
-     const val LITHO_VI: Double      = 7.0
-     const val ASTHENO_VI: Double    = 7.8
-     const val MANTLE_VI: Double     = 8.2
-     const val OUT_VI: Double        = 7.5
-     const val IN_VI: Double         = 12.0
+     const val CRUST_VI: Float      = 5.3f
+     const val LITHO_VI: Float      = 7.0f
+     const val ASTHENO_VI: Float    = 7.8f
+     const val MANTLE_VI: Float     = 8.2f
+     const val OUT_VI: Float        = 7.5f
+     const val IN_VI: Float         = 12.0f
 }
